refactor(core): rename logger module files for unique naming
