+++ packages/core/src/modules/logger/logger.config.ts
