+++ packages/core/src/modules/logger/logger.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '../../decorators/injectable';
 import { injectConfig } from '../../config';
 
-import { LoggerConfig } from './config';
+import { LoggerConfig } from './logger.config';
 
 type LogLevel = 'debug' | 'info' | 'warn' | 'error';
